provisioning changes
 - Separate Rabbit exchange/queue provisioning from the binder using the SPI provided in core
 - Refactor the common entities needed for Rabbit into a new core module
 - Refactor binder code to reflect the SPI changes

Fixes #40
