+++ pom.xml
@@ -51,6 +51,7 @@
 		</dependencies>
 	</dependencyManagement>
 	<modules>
+		<module>spring-cloud-stream-binder-rabbit-core</module>
 		<module>spring-cloud-stream-binder-rabbit</module>
 		<module>spring-cloud-starter-stream-rabbit</module>
 		<module>spring-cloud-stream-binder-rabbit-test-support</module>
